Register as sub-property of L.MarkerClusterGroup
instead of using local var, so that we comply with Leaflet plugin guideline.
https://github.com/Leaflet/Leaflet/blob/master/PLUGIN-GUIDE.md#plugin-api
It is also more consistent with `L.markerClusterGroup.layerSupport` factory.

Also got rid of "global" variables which slightly decrease access delay and file size / weight, but remain in memory for nothing, as they are used mainly during plugin initialization.