Refactor(debug/gh-pages): moved gh-pages into master/docs
so that it is easier to update the repository GitHub pages (just need to go into `docs/` folder, instead of having to switch to `gh-pages` branch).
It will also be possible to include the pages into the buildDocs process, so that the plugin version is updated automatically as well.
Also added .npmignore file.