Docs(debug/demos): use HTTPS for Tiles & OSM link
on all debug and demo/example pages (i.e. 6 HTML files), so that on `github.io`, there is no "mixed content" warning anymore.
Also updated the copyright link to OSM using the "official" URL (with https), even though it takes more bytes than the shortcut URL. But at least it is easier for visitors to understand the relationship between the link text and the URL.