Docs(demos): refactor throttling with manage-libs-versions
so that we can merge the demo and debug page, using either the CDN version or locally built version of the plugin.
Added text (expected result) and some debug printing, in order to make it clearer that we expect "nothing" visible and no error, otherwise we may think that it failed whereas it is actually the expected result!