Docs(demos): refactor HTML version selection table
so that it is dynamically built by "demo-setup.js" file and we can factorize its code for all demo pages. It is possible to customize the table (typically in order to add an extra column for another library), but to still have the demo-setup fill the cells for Leaflet, Leaflet.markercluster and Leaflet.MarkerCluster.LayerSupport.